Feature: BuyBox Size Swatches (#221)
This PR introduces a WIP Size Swatches "component".

• renders styled size options
• adds "active" CSS class when an option is selected
• updates the "Size" heading above the swatches with the currently selected size
• adds "Size Guide" link (does not do anything yet)

I do not expect the JS logic to be final. This was more to show how it might work. Once the color swatches and other BuyBox items start falling into place, we can reassess and take the JS logic where we need it to go.